niCURL/tsrc: Cleanup & update test URLs ;
